refactor: Revised import statement to reflect changes to the dependency.
